requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
Django==3.1.2
django-bootstrap4==2.3.1
djangorestframework==3.12.1
importlib-metadata==2.0.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.4.0
